requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
Markdown
packaging
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
setuptools
sqlparse
typing_extensions
tzdata
virtualenv
